Fix all ruff and mypy --strict warnings — closes #19, #20
lifecycle.py:
- Optional[X] → X | None throughout (UP045), remove unused Optional import
- noqa: B027 on intentional non-abstract signal_completion
- str() cast on _get_imds_token return (no-any-return)
- Shorten over-length docstring line (E501)

storage.py:
- type: ignore[import-untyped] for aioboto3 / botocore imports
- shard_bytes: tuple[bytes, ...] → list[bytes] (asyncio.gather returns list)
- zip(..., strict=True) on shard/checksum iteration (B905)

adapters/pyscf.py:
- int(...) cast on both checkpoint_size_estimate returns (no-any-return)
- ruff: noqa: RUF002, RUF003 for intentional Unicode math symbols

__init__.py: fix import sort (I001); noqa RUF022 on grouped __all__
cli.py: Annotated from typing (UP035); Optional to X | None (UP045)
protocol.py: fix import sort (I001)
pyproject.toml: mypy overrides for optional typer/rich deps

+++ src/spot_checkpoint/adapters/pyscf.py
@@ -1,3 +1,4 @@
+# ruff: noqa: RUF002, RUF003  (× is intentional Unicode math in docstrings/comments)
 """
 PySCF checkpoint adapters — SCF, CCSD, CASSCF.
 
@@ -123,7 +124,7 @@ class SCFCheckpointAdapter:
     def checkpoint_size_estimate(self) -> int:
         nao = self.mf.mol.nao
         # mo_coeff (nao × nao) + mo_occ (nao) + mo_energy (nao), all float64
-        return nao * nao * 8 + nao * 8 * 2
+        return int(nao * nao * 8 + nao * 8 * 2)
 
 
 class CCSDCheckpointAdapter:
@@ -284,4 +285,4 @@ class CASSCFCheckpointAdapter:
                 else 0
             )
 
-        return mo_bytes + ci_bytes
+        return int(mo_bytes + ci_bytes)
